third_party/WebKit/LayoutTests/fast/events/constructors/application-cache-error-event-constructor-expected.txt

This tests the constructor for the ApplicationCacheErrorEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new ApplicationCacheErrorEvent('eventType').bubbles is false
PASS new ApplicationCacheErrorEvent('eventType').cancelable is false
PASS new ApplicationCacheErrorEvent('eventType').reason is ""
PASS new ApplicationCacheErrorEvent('eventType').url is ""
PASS new ApplicationCacheErrorEvent('eventType').status is 0
PASS new ApplicationCacheErrorEvent('eventType').message is ""
PASS new ApplicationCacheErrorEvent('eventType', { bubbles: false }).bubbles is false
PASS new ApplicationCacheErrorEvent('eventType', { bubbles: true }).bubbles is true
PASS new ApplicationCacheErrorEvent('eventType', { cancelable: false }).cancelable is false
PASS new ApplicationCacheErrorEvent('eventType', { cancelable: true }).cancelable is true
PASS new ApplicationCacheErrorEvent('eventType', { reason: 'foo' }).reason is "foo"
PASS new ApplicationCacheErrorEvent('eventType', { url: 'bar' }).url is "bar"
PASS new ApplicationCacheErrorEvent('eventType', { status: 42 }).status is 42
PASS new ApplicationCacheErrorEvent('eventType', { message: 'baz' }).message is "baz"
PASS successfullyParsed is true

TEST COMPLETE
